refactor: generalises request match operators.
