--- core/api/src/main/java/io/gatehill/imposter/plugin/config/resource/ResourceMatchOperator.kt
@@ -1,5 +1,0 @@
-package io.gatehill.imposter.plugin.config.resource
-
-enum class ResourceMatchOperator {
-    EqualTo, NotEqualTo, Contains, NotContains, Matches, NotMatches, Exists, NotExists
-}
