+++ core/api/src/main/java/io/gatehill/imposter/plugin/config/resource/conditional/BodyMatchOperator.kt
@@ -1,0 +1,8 @@
+package io.gatehill.imposter.plugin.config.resource.conditional
+
+/**
+ * @author Pete Cornish
+ */
+enum class BodyMatchOperator {
+    EqualTo, NotEqualTo, Contains, NotContains, Matches, NotMatches, Exists, NotExists
+}
