+++ core/api/src/main/java/io/gatehill/imposter/plugin/config/resource/request/PathParamsResourceConfig.kt
@@ -40,17 +40,11 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with Imposter.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.gatehill.imposter.plugin.config.resource
+package io.gatehill.imposter.plugin.config.resource.request
 
 /**
- * Deprecated and superseded by [QueryParamsResourceConfig.getQueryParams].
- *
  * @author Pete Cornish
  */
-interface LegacyQueryParamsResourceConfig {
-    /**
-     * Use [QueryParamsResourceConfig.queryParams] instead.
-     */
-    @get:Deprecated("Use queryParams instead")
-    val params: Map<String, String>?
+interface PathParamsResourceConfig {
+    val pathParams: Map<String, String>?
 }
