+++ core/api/src/main/java/io/gatehill/imposter/plugin/config/resource/request/RequestBodyResourceConfig.kt
@@ -40,7 +40,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with Imposter.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.gatehill.imposter.plugin.config.resource.reqbody
+package io.gatehill.imposter.plugin.config.resource.request
 
 /**
  * @author Pete Cornish
