+++ core/api/src/main/java/io/gatehill/imposter/plugin/config/resource/request/RequestHeadersResourceConfig.kt
@@ -40,11 +40,13 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with Imposter.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.gatehill.imposter.plugin.config.resource
+package io.gatehill.imposter.plugin.config.resource.request
+
+import io.gatehill.imposter.plugin.config.resource.conditional.ConditionalNameValuePair
 
 /**
  * @author Pete Cornish
  */
-interface PathParamsResourceConfig {
-    val pathParams: Map<String, String>?
+interface RequestHeadersResourceConfig {
+    val requestHeaders: Map<String, ConditionalNameValuePair>?
 }
